backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
shotstack-sdk
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
